Various warning fixes

+++ modules/pvm/pvmvolumewriter.cpp
@@ -52,32 +52,34 @@ PVMVolumeWriter* PVMVolumeWriter::clone() const {
     return new PVMVolumeWriter(*this);
 }
 
-void PVMVolumeWriter::writeData(const Volume* data, const std::string filePath) const {
-    if (filesystem::fileExists(filePath)  && !overwrite_)
-        throw DataWriterException("Error: Output file: " + filePath + " already exists", IvwContext);
+void PVMVolumeWriter::writeData(const Volume *data, const std::string filePath) const {
+    if (filesystem::fileExists(filePath) && !overwrite_)
+        throw DataWriterException("Error: Output file: " + filePath + " already exists",
+                                  IvwContext);
 
-    const DataFormatBase* format = data->getDataFormat();
+    const DataFormatBase *format = data->getDataFormat();
     int components = 0;
 
-    #include <warn/push>
-    #include <warn/ignore/switch-enum>
-    switch (format->getId())
-    {
-    case inviwo::DataFormatEnums::UINT8:
-        components = 1;
-        break;
-    case inviwo::DataFormatEnums::UINT16:
-        components = 2;
-        break;
-    default:
-        break;
+#include <warn/push>
+#include <warn/ignore/switch-enum>
+    switch (format->getId()) {
+        case inviwo::DataFormatEnums::UINT8:
+            components = 1;
+            break;
+        case inviwo::DataFormatEnums::UINT16:
+            components = 2;
+            break;
+        default:
+            break;
     }
-    #include <warn/pop>
+#include <warn/pop>
 
     if (components == 0)
-        throw DataWriterException("Error: Output format " + std::string(format->getString()) + " not support by PVM writer", IvwContext);
+        throw DataWriterException("Error: Output format " + std::string(format->getString()) +
+                                      " not support by PVM writer",
+                                  IvwContext);
 
-    const VolumeRAM* vr = data->getRepresentation<VolumeRAM>();
+    const VolumeRAM *vr = data->getRepresentation<VolumeRAM>();
     const unsigned char *dataPtr = (const unsigned char *)vr->getData();
 
     size3_t dim = vr->getDimensions();
@@ -88,51 +90,50 @@ void PVMVolumeWriter::writeData(const Volume* data, const std::string filePath) 
     spacing.z = basis[2][2] / dim.z;
 
     unsigned char *data2Ptr = nullptr;
-    if (components == 2){
-        size_t size = dim.x*dim.y*dim.z;
-        data2Ptr = new unsigned char[size*components];
+    if (components == 2) {
+        size_t size = dim.x * dim.y * dim.z;
+        data2Ptr = new unsigned char[size * components];
         size_t bytes = format->getSize();
         memcpy(data2Ptr, dataPtr, size * bytes);
-        swapbytes(data2Ptr, static_cast<unsigned int>(size*bytes));
+        swapbytes(data2Ptr, static_cast<unsigned int>(size * bytes));
         dataPtr = (const unsigned char *)data2Ptr;
     }
 
     unsigned char *description = nullptr;
-    StringMetaData* descMetaData = data->getMetaData<StringMetaData>("description");
-    if (descMetaData){
+    StringMetaData *descMetaData = data->getMetaData<StringMetaData>("description");
+    if (descMetaData) {
         description = new unsigned char[descMetaData->get().size() + 1];
         strncpy((char *)description, descMetaData->get().c_str(), descMetaData->get().size());
         description[descMetaData->get().size()] = '\0';
     }
 
     unsigned char *courtesy = nullptr;
-    StringMetaData* courMetaData = data->getMetaData<StringMetaData>("courtesy");
-    if (courMetaData){
+    StringMetaData *courMetaData = data->getMetaData<StringMetaData>("courtesy");
+    if (courMetaData) {
         courtesy = new unsigned char[courMetaData->get().size() + 1];
         strncpy((char *)courtesy, courMetaData->get().c_str(), courMetaData->get().size());
         courtesy[courMetaData->get().size()] = '\0';
     }
 
     unsigned char *parameter = nullptr;
-    StringMetaData* paraMetaData = data->getMetaData<StringMetaData>("parameter");
-    if (paraMetaData){
+    StringMetaData *paraMetaData = data->getMetaData<StringMetaData>("parameter");
+    if (paraMetaData) {
         parameter = new unsigned char[paraMetaData->get().size() + 1];
         strncpy((char *)parameter, paraMetaData->get().c_str(), paraMetaData->get().size());
         parameter[paraMetaData->get().size()] = '\0';
     }
 
     unsigned char *comment = nullptr;
-    StringMetaData* commMetaData = data->getMetaData<StringMetaData>("comment");
-    if (commMetaData){
+    StringMetaData *commMetaData = data->getMetaData<StringMetaData>("comment");
+    if (commMetaData) {
         comment = new unsigned char[commMetaData->get().size() + 1];
         strncpy((char *)comment, commMetaData->get().c_str(), commMetaData->get().size());
         comment[commMetaData->get().size()] = '\0';
     }
 
-
-    writePVMvolume(filePath.c_str(), dataPtr, dim.x, dim.y, dim.z, 
-        components, spacing.x, spacing.y, spacing.z,
-        description, courtesy, parameter, comment);
+    writePVMvolume(filePath.c_str(), dataPtr, static_cast<unsigned int>(dim.x),
+                   static_cast<unsigned int>(dim.y), static_cast<unsigned int>(dim.z), components,
+                   spacing.x, spacing.y, spacing.z, description, courtesy, parameter, comment);
 
     delete[] data2Ptr;
     delete[] description;
